Moved source to nested folder in order to achieve better include paths. Updated CMake configuration.
