+++ src/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_library(rectpack2D INTERFACE)
+
+target_include_directories(
+    rectpack2D
+    INTERFACE
+        ${CMAKE_CURRENT_SOURCE_DIR}/
+)
