+++ src/rectpack2D/empty_spaces.h
