+++ src/rectpack2D/rect_structs.h
